+++ main/CenterAid.h
@@ -11,8 +11,8 @@
 #include "CircleWind.h"
 
 #define CA_NUM_DIRS 24  // every 15°
-#define CA_STEP (360.0/CA_NUM_DIRS) // 15
-#define CA_STEP_2 (CA_STEP/2.0)   // 30
+#define CA_STEP ((float)(360.0/CA_NUM_DIRS)) // 15
+#define CA_STEP_2 (CA_STEP/2.0f)   // 30
 #define MAX_DISK_RAD 6
 #define PEAK_STORAGE 120
 #define DRAW_SCALE (PEAK_STORAGE/MAX_DISK_RAD)
